Add server info pane and slow mode
The info pane gives the user a better idea of what
the server is doing at any given time. If printing
is required for debugging, it can be disabled with
the -n switch.

Slow mode puts a .5 second break in the server event
loop to make it easier to test client disconnects on
short jobs.